% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JEL.R
\name{JEL}
\alias{JEL}
\title{Jackknife Empirical log-likelihood ratio}
\usage{
JEL(X, theta, kerh)
}
\arguments{
\item{X}{data set, matrix or vector, if it is matrix, each row represents one sample.}

\item{theta}{the parameter needed to evaluated at.}

\item{kerh}{the kernel of the U-statistic,2-dimension only.}
}
\value{
The Jackknife Empirical log-likelihood ratio
}
\description{
Computation on Jackknife Empirical log-likelihood ratio
}
\examples{
\dontrun{
a <- rchisq(100,1)
kerh<-function(x,y) max(x,y)/2
JEL(X=a,theta = 0.83, kerh = kerh)
}
}
