% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Metr}
\alias{Metr}
\title{A random walk Metropolis sampler  using Rcpp}
\usage{
Metr(sigma, N)
}
\arguments{
\item{sigma}{the standerror of norm distribution}

\item{N}{the number of samples}
}
\value{
a random sample of size N
}
\description{
a random walk Metropolis sampler  using Rcpp
}
\examples{
\dontrun{
rnC <- Metr(1,100)
}
}
