% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/woodel.R
\name{woodel}
\alias{woodel}
\title{Log-likelihood ratio using sequential linearization method}
\usage{
woodel(X, theta, kerh)
}
\arguments{
\item{X}{data set, matrix or vector, if it is matrix, each row represents one sample.}

\item{theta}{the parameter needed to evaluated at.}

\item{kerh}{the kernel of the U-statistic,2-dimension only.}
}
\value{
Log-likelihood ratio using sequential linearization method
}
\description{
Computation on Log-likelihood ratio using sequential linearization method
}
\examples{
\dontrun{
a <- rchisq(100,1)
kerh<-function(x,y) max(x,y)/2
woodel(X=a,theta = 0.83, kerh = kerh)
}
}
